requirements.txt
googletrans
gtts
markovify
youtube-dl
ipython
google-cloud-texttospeech
discord.py[voice]
